revocation: Rename `types` module to `crl`.
